Re-use code for cancel/speed-up, add --gas-premium override
