+++ cmd/tx_list_pending.go
@@ -1,5 +1,5 @@
 /*
-Copyright © 2023 Glif LTD
+Copyright © 2025 Glif LTD
 */
 package cmd
